Refactor debug_hook()
This moves debug_hook() into its own file, eliminating the duplicate
implementation in fallback, and splits its responsibilities with a new
function, debug_wait().  debug_hook() continues to do the printing of
add-symbol-file and testing for the UEFI debug variable, and
debug_wait() handles the actual waiting for the debugger to attach.

Signed-off-by: Peter Jones <pjones@redhat.com>